Add files via upload
In this project, you will write most of the code yourself (you should be able to do so by taking reference from whatever has been taught in the course as well as the code we have shared).

Task 
You will implement a 3-class Image Classifier using a CNN on a dataset consisting of 1000 images each of Panda, Cat and Dog. The dataset is taken from Kaggle. We have split it into 80:20 ratio for train:validation which you can download from here.

You should build your own network and not use any pre-trained network/weights for this task. 

The purpose of this Project is two-fold:

The first is to get your hands dirty with training a neural network from scratch so that you understand what it takes to build and train a real network on real data. You will implement everything - right from getting the data to analysing the final results.

The second is to reinforce the use of the complete training pipeline we suggested in our previous lecture. In practice, You should try to stick to this pipeline whenever you work on a new problem.