Add files via upload
TorchScript is a way to create serializable and optimizable models from PyTorch code. Any TorchScript program can be saved from a Python process and loaded in a process where there is no Python dependency. For example, a model trained using PyTorch can be serialized and saved using TorchScript, and that serialized models can be loaded and used in the C++ code.